Move auto refresh to header.
