+++ frontend/src/components/cluster-dashboard/ClusterDashboard.scss
@@ -5,51 +5,3 @@
   margin-bottom: $footer-height !important;
   padding: 0px 0px $single-spacing 0px;
 }
-
-.auto-refresh-wrapper {
-  position: absolute;
-  top: 0px;
-  right: $double-spacing;
-  height: $header-height;
-  padding: $single-spacing 0px;
-
-  .MuiTypography-root {
-    vertical-align: middle;
-    font-size: $text-size-xs;
-    color: $secondary-theme-color;
-    margin-right: $single-spacing;
-    font-weight: bold;
-  }
-
-  .refresh-slider-label {
-    font-size: $text-size-xs;
-    color: $secondary-color;
-  }
-
-  .MuiSlider-root {
-    width: 100px;
-    top: 8px;
-    margin: 0px $single-spacing;
-
-    .MuiSlider-mark {
-      background-color: transparent;
-    }
-
-    .MuiSlider-rail {
-      background-color: $secondary-color;
-    }
-
-    .MuiSlider-thumb { 
-      background-color: $primary-color;
-    }
-
-    .MuiSlider-valueLabel {
-      top: -$double-spacing;
-
-      * {
-        background-color: transparent;
-        color: $secondary-theme-color;
-      }
-    }
-  }
-}
